refactor(dag:*post_insights): deduplicate common code with *PostsInsightsParser
